streamlit/requirements.txt
dash==2.0.0
plotly==5.0.0
psycopg2-binary==2.9.3
werkzeug==2.0.3
numpy==1.23.5  # Versi yang cocok dengan pandas
pandas==1.5.3  # Sesuaikan versi sesuai kebutuhan Anda
Flask
requests

